# Fix Duplicate Centroids Entries and Incorrect Heuristic Names in Tables