… CLEANUP_SUMMARY.md
… PLOTS_AND_TABLES_FIX_SUMMARY.md
… analysis_outputs/results/plots/comprehensive_metrics_comparison.csv
… analysis_outputs/results/plots/domain_analysis.csv
… analysis_outputs/results/plots/efficiency_analysis_table.csv
… analysis_outputs/results/plots/heuristic_summary.csv
… analysis_outputs/results/plots/performance_summary_table.csv
… analysis_outputs/results/plots/quality_analysis_table.csv
… analysis_outputs/results/plots/timing_analysis_table.csv